+++ GPipe-GLFW/src/Graphics/GPipe/Context/GLFW/Calls.hs
@@ -235,3 +235,8 @@ setDropCallback onMain window cb = onMain $ GLFW.setDropCallback window cb
 -- * This function must only be called from the main thread.
 getWindowSize :: OnMain (Int, Int) -> GLFW.Window -> IO (Int, Int )
 getWindowSize onMain window = onMain $ GLFW.getWindowSize window
+
+-- |
+-- * This function must only be called from the main thread.
+setWindowSizeCallback :: EffectMain -> GLFW.Window -> Maybe GLFW.WindowSizeCallback -> IO ()
+setWindowSizeCallback onMain window cb = onMain $ GLFW.setWindowSizeCallback window cb