+++ GPipe-GLFW/CHANGELOG.md
@@ -1,6 +1,11 @@
-### UNRELEASED
+### 2.0.0 -- UNRELEASED
 
-* add `setWindowSizeCallback`
+* Add `setWindowSizeCallback` so that it is no longer necessary to poll for
+  window size changes.
+* **Potentially breaking change:** Add `configEnableDebug :: Bool` field to the
+  `ContextHandlerParameters Handler`/`HandleConfig` data constructor.
+* Suppress debug logging by default. Turn it back on with
+  `defaultHandleConfig{configEnableDebug = True}` config.
 
 ### 1.4.1.4
 
@@ -13,7 +18,7 @@
 
 ### 1.4.1.2
 
-* [Kludgy](https://github.com/LinuxUser404) bumped versions and resolver.
+* [Kludgy](https://github.com/Kludgy) bumped versions and resolver.
 * [LinuxUser404](https://github.com/LinuxUser404) bumped versions and added `cabal.project` & `.travis.yml`.
 * Set up travis builds with stack.